Move generate-feed to webpack plugin
Instead of having generate-feed as separate step, take advantage of
webpack and generate the feed from template and data constructed during
webpack build.
